Move to a dedicated component.
