+++ viz/app/components/VizContext.tsx
@@ -4,10 +4,18 @@ import { createContext, useContext } from "react";
 
 interface VizContextValue {
   isPdfMode: boolean;
+  editText:
+    | ((
+        editId: string,
+        oldText: string,
+        newText: string
+      ) => Promise<{ success: boolean }>)
+    | null;
 }
 
 export const VizContext = createContext<VizContextValue>({
   isPdfMode: false,
+  editText: null,
 });
 
 export function useVizContext(): VizContextValue {
